Move table_manager functions into schema module
Consolidates get_project_db_engine, create_tables, and evolve_table
from table_manager.py into schema.py. Updates all imports.
