--- corehq/apps/project_db/table_manager.py
@@ -1,48 +1,0 @@
-from alembic.migration import MigrationContext
-from alembic.operations import Operations
-from sqlalchemy import inspect
-
-from corehq.sql_db.connections import DEFAULT_ENGINE_ID, connection_manager
-
-
-def get_project_db_engine():
-    """Return a SQLAlchemy engine for project DB tables.
-
-    Currently uses the default Django database. This will later switch
-    to a dedicated ``'project_db'`` engine ID.
-    """
-    return connection_manager.get_engine(DEFAULT_ENGINE_ID)
-
-
-def create_tables(engine, metadata):
-    """Create all tables in ``metadata`` that don't yet exist.
-
-    Uses ``checkfirst=True`` so existing tables are left untouched.
-    """
-    metadata.create_all(bind=engine, checkfirst=True)
-
-
-def evolve_table(engine, table):
-    """Add columns and indexes present in ``table`` but missing from the database.
-
-    This is append-only: columns and indexes that exist in the database
-    but not in ``table`` are left in place (never dropped).
-    """
-    inspector = inspect(engine)
-    existing_columns = {col['name'] for col in inspector.get_columns(table.name)}
-    new_columns = [col for col in table.columns if col.name not in existing_columns]
-
-    existing_indexes = {idx['name'] for idx in inspector.get_indexes(table.name)}
-    new_indexes = [idx for idx in table.indexes if idx.name not in existing_indexes]
-
-    if not new_columns and not new_indexes:
-        return
-
-    with engine.begin() as conn:
-        op = Operations(MigrationContext.configure(conn))
-        for column in new_columns:
-            col = column.copy()
-            col.table = None
-            op.add_column(table.name, col)
-        for index in new_indexes:
-            index.create(bind=conn)
